[dev][docs] format markdown files
Excellent documentation is an explicit goal for both inline and
accompanying materials. This patch adds Prettier formatting for markdown
files to keep the latter as consistent as possible without any manual
intervention. There is no other standard tooling in the Wikimedia
ecosystem available for this purpose so far as I know.