[dev] Disable Prettier for everything except JSON and Markdown files
T256693 added a Prettier configuration file to Wikimediafy the
formatting. Unfortunately, some tooling assumes that if a Prettier
configuration is present, it applies to all files. Since we currently
only use Prettier for JSON and Markdown, this means it incorrectly runs
on everything else. This patch updates the Prettier ignores to exclude
those files.