[docs][dev] Document source maps
I've used source-map-explorer before and it gives a pretty nice
breakdown of style and program maps but isn't perfect. I haven't used
webpack-bundle-analyzer much, but it seems to do about the same thing,
and integrates better with Webpack. Let's try them both for this project
and remove what we don't want later.