[dev] Improve ESLint TypeScript config and fix JS config typing (#37)
eslint-plugin-jsdoc has a TypeScript mode. Enable it. Unfortunately, it
doesn't seem to work properly at the moment so disable a couple rules it
says it's supposed to that are causing errors. A similar approach was
taken recently [in the Popups repo] by one of the linter maintainers.

Also, fix the typing in the Storybook and Webpack configs. These aren't
currently being evaluated by the type checker as Webpack shakes them
out as non-functional. However, they can now be tested manually with
`npx tsc --noEmit --incremental false`. The readme has been revised.

[in the Popups repo]: https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/extensions/Popups/+/536470c01d5bc1e798d9790f2df5645736b8ddcd%5E%21/#F0

Co-authored-by: Anri Bolon <66359773+speedandfunction-anribolon@users.noreply.github.com>