[build][dev] Enable development releases
Decouple versioning from publishing to allow arguments to be passed to
`npm publish`. This is needed to allow pushing the `@next` NPM tag via
`--tag next`.

I've tried to test out the documented procedures locally in a dummy repo
but it's possible I might need to give it another few iterations once
this is merged.