[update] Add initial icon component
This basic icon component can be used by itself or within another
component (e.g Button, Message) but currently uses OOUI classes
for the icon background image. Once we figure out how we want to
replicate these styles and include the images we can ditch the
OOUI class names, but for now this will only work if those
styles and images are available.

Image inversion is covered, progress is not.