[build][fix] Don't push tags for development releases
Development releases are reproducible by passing the timestamp to the
build script. Ideally, we'd keep publishing these tags so they could be
built directly from the commit. However, this is creating a lot of noise
and is unusual. Limit tag publishing to only alpha / beta / release
candidate and final releases.