[dev] Add icon objects and icon groups
The Icon component will be updated to display an SVG rather than
using a background-image. This commit adds a file with the SVG
path (and other data) for each Wikimedia icon, plus a file with
icon group objects for use in Storybook and by developers who
want to import a subset of icons rather than the entire library.