[dev] Convert Icon component to SVG icon
Instead of using CSS classes and background images to display icon
images, use the <svg> element. This has several benefits:

- Makes it easy for developers to specify color and size
- Enables our Icon component to display custom or third-party icons
- Allows developers to import only the icons they need

This iteration includes a file with all icons and a file with icon
groups (as previously specified in OOUI). Developers can import
single icons or groups of icons.

A `flippable` property in the icon object tells the Icon component
whether it should use CSS to flip the icon when text direction is
rtl. Some Wikimedia icons can't just be flipped in this way and
require different SVG paths per text direction, which is supported.

Some Wikimedia icons require different SVG paths per langauge (e.g.
bold and italic). This is supported if a language map is defined
along with the different paths.