[component][typeahead-suggestion] Add network client for typeahead-search component (#72)
* Add fetch wrapper

In most browsers we can use window.fetch. In others, like IE 11, this is not available.
To keep this simple for now reject the promise if native fetch is not available.

Added jest-fetch-mock for testing the "fetch is available" branch.

* Add buildQueryString()

A convenience method to serialize a JS object into a query string, so it can be appended
to the URL (after '?').

* Add REST API search client

This network client provides a fetchByTitle() function, to be used for typeahead search suggestions.
Not providing accept-language header at this time since it does not seem to make a difference.

* Allow API tests to run against live servers

By default the requests are mocked, so no network is used.
If the environment variable TEST_LIVE_REQUESTS is set to
a truthy value, the mocks are disabled and requests are made
against a live server.

* Remove excerpt

A repeat of `title` doesn't seem useful to me.

* Move jestFetchMock.enableFetchMocks() into beforeAll

Also undo in afterAll().
Plus a formatting fix.

* Remove explicit typing of local var

* Mark width and height as nullable

* Remove explicit typing of local var

* Explicitly type RestResponse in test

Doing that I noticed that the null values need to be called out in the interface.

* Remove excerpt from RestResponse, too

* Add searchText to the search response

so it can be used by the UI for highlighting the search text.

* Avoid sending a request if there is no searchText

* Trim search text

Titles don't start or end with spaces. So, remove them, at least for title search.

* Update test name

* Handle null thumbnail width/height

The REST API documentation says it could return null for width or height. Make the value undefined in this case.

* Reuse input object for test verification

Clarify that the input object changes in one place where the thumbnail is null.

* Replace mockResponse with mockOnce

* Check that fetch was not called when searchText has only a space

* Add network error test

* Prefer toStrictEqual over toEqual

* Rename variable names searchText with query

* Rename restApiSearchClient.ts to restSearchClient.ts