[update] Add initial icon component (#47)
* [update] Flesh out Button props

Add some of the component properties and tests needed by Button.

Bug: T256038

* [update] Add initial Button styles

The following seven rules appear to break lengthy rules, such as multi-property
transitions and mixins, that must appear across multiple lines:

- declaration-colon-space-after
- declaration-block-semicolon-newline-before
- declaration-block-semicolon-space-before
- value-list-comma-newline-after
- function-parentheses-space-inside
- function-comma-newline-after
- function-parentheses-newline-inside

What to do?

The presentation seems pretty close to the filled styles in the OOUI
button showcase.

Is the class name and cascade structure wanted?

Can we keep theming isolated in wikimedia.less for now and sort out full
theming later?

* [dev] Revise Wikimedia theme file structure

* [dev] Remove a redundant comment

* [dev] Fix touch area comment

* [docs] Add Less import notes and miscellaneous revisions

* [dev] Move PrimaryAction out of button

* Fix lint config

* [update] Add initial icon component

This basic icon component can be used by itself or within another
component (e.g Button, Message) but currently uses OOUI classes
for the icon background image. Once we figure out how we want to
replicate these styles and include the images we can ditch the
OOUI class names, but for now this will only work if those
styles and images are available.

Image inversion is covered, progress is not.

* [dev] Use slot for screen reader content

Icon text is meant only for screen readers and will usually be
simple text, but to account for the possibility of HTML we can
use a slot then visually hide its contents.

* [dev] Improve icon documentation and remove unneeded styles

Some legacy OOUI styles may not be needed, so let's not include
them unless they become necessary as we build out more components.

* [update][button] Use rootClasses pattern

* [dev] Add comments about icon size calculation

* [docs] Add playground links

* [dev] Remove unneeded linter overrides

* [dev] Add icon objects and icon groups

The Icon component will be updated to display an SVG rather than
using a background-image. This commit adds a file with the SVG
path (and other data) for each Wikimedia icon, plus a file with
icon group objects for use in Storybook and by developers who
want to import a subset of icons rather than the entire library.

* [dev] Convert Icon component to SVG icon

Instead of using CSS classes and background images to display icon
images, use the <svg> element. This has several benefits:

- Makes it easy for developers to specify color and size
- Enables our Icon component to display custom or third-party icons
- Allows developers to import only the icons they need

This iteration includes a file with all icons and a file with icon
groups (as previously specified in OOUI). Developers can import
single icons or groups of icons.

A `flippable` property in the icon object tells the Icon component
whether it should use CSS to flip the icon when text direction is
rtl. Some Wikimedia icons can't just be flipped in this way and
require different SVG paths per text direction, which is supported.

Some Wikimedia icons require different SVG paths per langauge (e.g.
bold and italic). This is supported if a language map is defined
along with the different paths.

* [dev] Add optional langCode prop

In some cases (e.g. a UI with multiple languages) we may need to
explicitly define the langCode for a particular icon rather than
inferring it from the document lang attribute.

* [dev] Ensure icons scale with font size

Rather than providing a size option, we'll set icons up to be
the em equivalent of 20px by 20px at our base font size, and
to scale with font size.

* [dev] Add icon types

Per PR feedback, this standardizes icon data a bit by adding
types and simplifying language and text dir mapping.

* [dev] Handle more special cases

- Handle more icons that vary per language
- Handle icons that should be flipped in RTL mode with some
language-specific exceptions (e.g. help icon should be flipped in
RTL mode, but not for Hebrew or Yiddish

* [dev] Use SVG title and remove slot

Instead of using a visually-hidden slot to provide text to screen
readers, use the SVG title element and an iconTitle prop. If the
prop isn't used, assume the SVG to be presentational and hide it
from the accessibility API via aria-hidden.

* [dev] Remove icon groups

* [dev] Use proposed class object naming convention

* [dev] Update flip class name

* [dev] Update icon types and reorganize code

- Simplify structure of IconVariedByDir
- Rename language map property of IconVariedByLang
- Pull icon attribute functions into types file
- Move types into Icon component directory
- Improve documentation

* [dev] Use computed style to determine text direction

* [dev] Don't export icons that are part of a set

* [dev] Add icons and Icon component to main entry

* [dev] Refactor icons with a single path

Icons with a single path can be simple strings unless they should
flip for RTL text direction. This commit changes the Icon type to
be an alias for string and adds a new type, IconFlipForRtl, to
represent icons that should flip.

* [dev] Increase bundle size

* [dev] Fix some RTL icons

* [dev] Improve icon type documentation

* [dev] Change default color, SVG attributes, and more

- Set default color to #000 instead of currentColor
- Use pixels for SVG height and width to match viewbox size
- Remove unneeded user-select rule

* [dev] Improve dead code documentation

Co-authored-by: Stephen Niedzielski <stephen@niedzielski.com>
Co-authored-by: Volker E <Volker-E@users.noreply.github.com>

Bug: [T256038](https://phabricator.wikimedia.org/T256038)