[search] Allow consumer fetch implementation to be passed
Expose the new REST search API client implementation for WVUI consumers.

The typeahead-search component will accept a search client which accepts
a fetch implementation. This allows for fetch to be polyfilled (by
jQuery for example) or, if unspecified uses the default, window.fetch().
However, the search client itself wasn't actually exposed previously so
clients had no way to use it previously.

The bundle size increases with this patch because a search component
that actually uses the search client implementation has not yet been
merged.

This change happened to trigger a false positive ES5 check on
`Object.assign()`. This is a known issue in Popups. The unminified code
in question is:

  var w = function () {
    return ( w =
      Object.assign ||
      function ( t ) {
        for ( var e, n = 1, i = arguments.length; n < i; n++ ) {
          for ( var u in ( e = arguments[ n ] ) ) {
            Object.prototype.hasOwnProperty.call( e, u ) && ( t[ u ] = e[ u ] );
          }
        }
        return t;
      } ).apply( this, arguments );
  };

`Object.assign()` is polyfilled.

Change-Id: Icda1a9ec1981ec560b282ecb391cc4d2ae113cd8