Blubber config npm publish
Get CI to create a dev release of WVUI on every merge into master,
and publish the new version to npm. The project already had
a release script to do the publishing, this just hooks it up
to PipelineLib.
After this is merged, a secondary patch will need to be merged in
the integration/config repo to ensure WVUI runs the
newly added 'release' pipeline

Bug: T264244
Change-Id: I6a6e6ecf49b547d6571ef69420a3c2264bb17d85