Delete actionSearchClient and associated tests
From my understanding, we are planning on using the rest search client
(not the action search client) with Vector. The action search client
adds a maintenace burden and with no concrete payoff at this time. If a
use case arises, this commit can be reverted through the power of git.

Bug: T244287
Change-Id: I6c89464c11c19166da0c5fb1b31d2438882b5f25