[components][typeahead-search][typeahead-suggestion] Add `showThumbnail`/`showDescription` props
There is a need to optionally suppress the rendering of the thumbnail
and/or description in the TypeaheadSuggestion/TypeaheadSearch
components. This adds two props to TypeaheadSearch that are then passed
TypeaheadSuggestion which enable this ability.

Storybook knobs were also updated/created for easier testing of these
props.

Appropriate styling of these components when these two props are
`false` is deferred until T263657.

Additional changes:

* TypeaheadSuggestion snapshots were added to/updated

Bug: T260167
Change-Id: I8c8fef266266a21d0cf2577a60ac89a51aa565b5