[components][typeahead-search] Tidy onInput method definition
The previous syntax was used because the method was wrapped in a call to
lodash.debounce. That call was removed in PS23 of the patch that
introduced it (I50715e54) but the definition was never tidied.

The current syntax is desirable as it as the "this" keyword is now
correctly typed.

Change-Id: I98d8db94b675a7c77edf85d8fda3e78dcc491c13