Add Blubber generated Dockerfile to gitignore
In the post-merge step in Jenkins, git status
is run to ensure no uncommited changes exist.
But the Blubber generated Dockerfile is
written to the .pipeline directory during
one of the pipeline jobs, making the
git status fail as there are un-commited
file changes. This adds the generated Dockerfile
to .gitignore so its ignored by
git status.

Bug: T264244
Change-Id: I86c9c1039af304b21fbe4a7a9a7953035afbfccc