[typeahead-search][typeahead-suggestion] Unify figure size and spacing between figure and text
This replaces the padding-inline-start/padding fallback styles that were
previously applied to the suggestion and footer text in favor of
applying a single margin-right style on the suggestion thumb and footer
icon. Both of these approaches achieve the same goal of adding space
between the thumbnail/icon and text, but the latter is much less tedious
to write.

I believe the previous styles were created in an effort to support RTL
within WVUI without a reliance on Resource Loader. However, from my
understanding, the topic of RTL within WVUI is now a much larger
discussion that is expected to take place later. For now, the simplest
solution is to use Resource Loader's RTL flipping until a more robust
solution is determined.

Additionally:

* A `margin-end-typeahead-suggestion-thumb` variable was added to
wikimedia-ui.less which are consumed by both TypehaheadSearch and
TypeaheadSuggestion components for consistent spacing.

* The width of the input icon was set to
`@min-width-typeahead-suggestion-thumb` which is also used for the
typeahead suggestion thumb width and the footer icon width.
Additionally, the translate: -12px style was replaced with a more
calculated value.

Bug: T263657
Change-Id: I0d4530bf54ee2669d8b089f912f2d65bab6e33f2