[components][typeahead-search][typeahead-suggestion] Style `showThumbnail` prop
Follow-up to 6dc2e3d0616dbadc6f8c5990b5e41eface6be51b which styles the
the `showThumbnail` prop according to the spec in T263657.

This adds a block modifier `wvui-typeahead-search--show-thumbnail` class
to typeahead search's root element which is acceptable according to BEM
guidelines [1]. In addition, the spacing and size of the the
input icon container, suggestion thumbnail, and footer icon container
are unified through variables so that these elements vertically line up
in an aesthetically pleasing manner.

[1] http://getbem.com/faq/#block-modifier-affects-elements

Bug: T263657
Change-Id: Id92354c981bcc927b6e0c74551f210a6ef6167c9