[components][icon] Remove `font-size` property
`font-size` property was intended to provide a user-settings zoomable
ability to resize icons. This promise has never been fulfilled.
SVG width is set in pixels and in order to be overridable, `svg`
width and height need to be overriden with a relative unit value.
`font-size` as provided is also blocking clear icon overrides (reducing
width) as intended in input component.

Bug: T256040
Change-Id: I199dd0a44d923b3a87af81af1500da9725df6cc4