[components][typeahead-search] Further tidy onInput method definition
Extract the updateSuggestions and clearSuggestions methods in order to
tidy up the onInput method and ease reasoning about its happy and sad
paths.

Moreover, Jdrewniak points out that the async keyword can't be used in
the definition of the onInput method as it transpiles to native
promises, which aren't supported by IE11. Leverage the promise returned
by SearchClient#fetchByTitle to avoid this situation.

This change supersedes I98d8db94.

Bug: T257698
Change-Id: I42e38fa7298525f6ec4ad64e0c07c49ab1413c9e