Publish new wvui development versions through Docker container
This patch is to fix /bin/release-dev so you can
publish development releases through a Docker container.

This patch introduces a new docker compose-service (release)
through which releases should now be done. It also requires
a new set of ssh keys specifically for doing releases,
as well as modifications to the release script to ensure
an npm token has been generated for successfully publishing the
new release to npm. This should just fix development releases
for now, will make necessary changes to release-pre and
release-prod scripts in a separate patch.

The project README has been updated as well to document all
the above changes.

Bug: T263787
Change-Id: Ia567c538363652736b782884bace8116774844e5