[components][typeahead-search] Show footer when there aren't suggestions
I42e38fa7 introduced a bug where the footer suggestion would be hidden
when the search backend didn't return a suggestion.

Update updateSuggestions to keep the suggestion list visible
("expanded") if the user is searching for something. This mirrors the
logic in onFocus. Accordingly, update onSuggestionClick to explicitly
hide the suggestion list.

Bug: T257698
Change-Id: I4781c54f65eaff84079aea54301a0704ccbc1dbe