[typeahead-search] Correct suggestions width
This is effectively a revert of
I0857ec86457b3cf2a20c54a7d89f42264c250802 to correct the suggestions
width. According to the design spec T263657, the suggestions list should
not pass the button.

The one notable downside of this approach (as mentioned by Volker in his
commit) is that it adds an extra tab to get to the search button from
the input.

Change-Id: Ifefa2bc813e2c49cc00ce8c064344802bf985080