[typeahead-suggestion] Replace min-width with flex-shrink
Using a min-width property here with a variable makes this style more
complicated than it needs to be if using `flex-shrink` can be used
instead. In both cases, the intended goal is to prevent the thumbnail
from shriking when the descriptions are too wide.

In a future patch I hope to get rid of the
`@min-width-typeahead-suggestion-thumb` variable altogether.

Bug: T263657
Change-Id: I56a03a674c79e38b95bc54df8f45cb3ef32d0a36