[components] Use `@size-search-figure` variable across Input, TypeaheadSuggestion, TypeaheadSearch
This cleans up tech debt by making the figure widths used in the input,
typeahead suggestion, and typeahead search components consistent.

It drops the use of the `@size-typeahead-suggestion-thumb` and
`@width-typeahead-search-figure` variables in favor of using a new
`@size-search-figure` variable that is used across these three
components. We do this because when these three components are used
together (in the typeahead search component), the figures need to
vertically line up nicely and this limits the amount of style overriding
that needs to happen in order to do this.

Change-Id: Id7d7aa1e951dd75cb8e3e98b2160aaa268224f34