[typeahead-search] Make non-blank `initialInputValue` prop call `onInput`
In Vector, a user can type characters into the input before WVUI loads.
To account for this, we have the `initialInputValue` prop which sets the
WVUI's initial input value. However, passing this prop wasn't triggering
the `input` event and therefore search results would not show when the
component was mounted.

To get around that, during a mount a manual call to the onInput method
can be performed.

Change-Id: I45e061fc37886afece45887754fb2c1e3e355453