[typeahead-search][input] Correct spacing around input start icon and end icon
According to T270202, there should be equal space around the input's
start icon (or end icon). Note with the current state of input's inset
border, this measurement depends on whether the input is focused or not
and is thus somewhat ambiguous.

Because we place the start icon/end icon relative to the input's
container, we need to account for the size of the input's border. This
commit accomplishes this by:

* Replacing `@start-input-text-icon`/`@end-input-text-icon` variables
with a single `@size-input-icon-container` variable that determines the
total width of the icon container (including padding). This simplifies
input padding calculations.

Additional changes:

* Removes unnecessary "clearable" styles that were redundant and are
already accounted for by the preceeding changes.

* Changes TypeaheadSearch component to reflect these changes and fix the
same spacing issues.

Bug: T270202
Change-Id: I0ec3dcedaea90b01fe94e3416ee68ea33b782b4b