Install NPM from buster backports
This patch installs npm and node-ssri from buster backports for both
local development and in Jenkins CI. The reason for installing from
backports is a need for a higher version of npm that allows a certain
flag to be passed to the release scripts. The current npm version in
backports is 7.4.0. This patch also adds node-gyp 7.1.2 to the npm
packages because a higher version is required than can be installed
with apt.

Bug: T269762
Bug: T269957
Change-Id: Ib5e1cf577e51dc68133d2b1476df44774e8888af