[docs] Add contributing guidelines & refactor the existing README
- Introduce CONTRIBUTING.md, which includes information on contributing
via communication, patch authorship, and patch review. These guidelines
are a baseline and may require updates or additions as the project
progresses.
- Remove the bulk of the README and move it to a DEVELOPERS.md file.
This has a couple of advantages: the README is more focused on end
users, the lengthy table of contents is broken into two parts and most
of it is hidden from end users who don't need to see it, and the
heading hierarchy in the DEVELOPERS.md file loses one level of
complexity.
- The only update made to the content of DEVELOPERS.md was to fix some
formatting and to update the table of contents, no content was added
or changed.

Bug: T272883
Change-Id: Ie0ca0b787dac9724d62cbfd29d87036222bc117a