[components][button] Add 'type' prop to replace 'quiet', add 'primary'
Replace the boolean 'quiet' prop with a 'type' prop that takes three
values: 'normal', 'primary' or 'quiet'. Primary buttons are new to
WVUI, and follow the design specified in the design style guide.
Normal buttons follow the existing framed (non-quiet) style, and
quiet buttons follow the existing quiet style.

Change the CSS class naming accordingly.

Rotate the table in the "All Combinations" button story so it's not as
wide, and fits better on narrow screens.

Bug: T280840
Change-Id: I09380759d5122aced67175717f9ce9cd20aaeb1b