[icon] Remove iconColor prop, control color with CSS instead
This removes the iconColor prop from the Icon component, and instead
controls the color of the icon by setting the text color in CSS.

This allows us to control icon colors directly rather than by setting
opacities, so also add a @color-accessory variable that's equivalent to
@opacity-icon-accessory. This was upstreamed to wikimedia-ui-base in
Ia91f24a3, and should be removed once we upgrade.

Set the icon color to Base10 by default, and make components that embed
icons override this rule if they want to change their color. This occurs
in the Input, TypeaheadSearch and TypeaheadSuggestion components (where
opacities were being used), and in the Button component (where the text
color varies, and we want the icon color to match).

Bug: T280934
Change-Id: I8e583a3565a5dd7d4d01e8e78fe6338d97994de8