[build] Create a dedicated "wvui-search" entry point and bundle
Adds a new wvui-search entry point to src/entries. For the time
being this is identical with the main "wvui" entry point. However,
over time it is expected that these will diverge as new components
get added to the library which are not needed for the Vector search
feature.

The wvui-search entry point produces the following output:

dist/commonjs2/
├── wvui-search.commonjs2.js    # unminified JS files
└── wvui-search.css             # unminified CSS

Going forward, "wvui" will be the main entry point for the library
(provided in both UMD and CommonJS formats for use both inside and
outside of MediaWiki). wvui-search will *only* be provided in
CommonJS format, since it is only intended for use within the MW
environment. The wvui-icons endpoint is left unchanged.

Bug: T280828
Change-Id: I54da82fdd70ef9cce7df96e0f376b05ef9b4869c