[checkbox] Add the Checkbox component
Adds a basic checkbox component and associated Less variables.

Like Radio, the Checkbox component is meant to be used with v-model
to bind a value for a single checkbox or group of checkboxes, and
to emit an event on change. For a single checkbox, the value bound
to v-model will be boolean; for a group of checkboxes, it will be
an array (of strings or numbers matching the individual input values
of all "on" checkboxes in the group).

An indeterminate prop can be provided to the component to set the
`indeterminate` property on the HTMLInputElement object. When this
property is set to true, the visible checkbox span will display an
indeterminate style, rather than checked or not checked. However,
the logic to set the indeterminate prop for a checkbox must be
handled in the parent component by the library user (e.g. in the case
of nested checkboxes).

Bug: T281191
Change-Id: I0bdce8ee007fc314ae50a6b30b29b6f9789073d9