[radio] Improve handling of modelValue prop in Radio stories
Radio has a prop called modelValue, which is provided by v-model,
and whose custom prop name is set via the `model` option in the
component. Storybook ends up calling this prop 'v-model', which
is confusing since v-model is not a prop, it's the provider of the
modelValue prop. To remedy this, we need to provide the modelValue
name in the argType for 'v-model'.

Additionally, this patch swaps out an em value for px in the
Storybook-specific Radio styles.

Change-Id: I77221c4c3cd827e544f2aa020620a8ae2dbdf056