[radio] Remove superfluous `aria-disabled` and restructure disabled rules
`aria-disabled` is already set on wrapper `label`, with it all focusable
descendants are disabled according to WCAG and the extra attribute is
superfluous.
It has been a good practice to put disabled rules at the end of selector
and group enabled and disabled together in this order. While there move
`:hover` styles only to enabled selector as disabled components should
generally not react to user interactions.
Also `inherit` cursor to save another rule.

Change-Id: Idc72aa9c0f9b01848f2b4896bf498e75900ac9f1