[binary inputs] Add layout styles and options for Checkbox and Radio
Most checkboxes and radios that appear in groups display in a stacked
layout. This patch adds styles for that layout by default, plus an
`inline` prop that, when set to true, will set inline styles instead
via a conditional CSS class.

Note that the `inline` prop is not configurable in the stories for
these components because there is little reason to toggle between
stacked and inline styles for a single group of inputs: inline
styles should only be used in very specific cases (2-3 inputs with
very short labels), and other cases are not covered by the styles
(e.g. extensive handling of inline inputs spanning multiple lines).

Bug: T284337
Change-Id: Ia067418bb902e2c7e0de947afb1d4dd1469017f3