[types] Move type declarations into "types" dir, exclude unneeded files
Summary

* Updates tsconfig so that type declarations are output to a new
  "types" dir instead of "dist"
* Updates package.json to include this folder in the NPM package
  and points the "types" key to this new location
* Adds an .npmignore inside types/ to ensure that only declaration
  files for actual library code get included in the NPM package.
  Declaration files for tests, stories, etc. should not be shipped.

Notes

Moving the types into a dedicated folder should make them a bit easier
for consumers to discover if they need to reach in and dig for things
that are not surfaced in the main entry point (for exmaple, prop types
like ButtonAction).

Types are still automatically generated from tsc and are not tracked
in git. However, if we do decide later that we want to manually maintain
these files because of limitations in the generated output, we could
start tracking this folder and stop emitting the declarations on
every build.

Bug: T284782
Change-Id: Icb64713204d59f05c519d41a9bfd71fbfc7f8b6d