[ToggleButton] Add new toggle button component
Styling is based on existing OOUI widget styles. Decided not
to wrap around a wvui-button since the colors would need to
be different and we don't care about the different actions and
types, so the base styles for all versions of the toggle button
(disabled or not, active or not) are based on the ones for the base
version of wvui-button.

Requires bumping the size of the "wvui.js" and "wvui.css"
distribution bundles by .1 KB each.

Bug: T287135
Change-Id: I6b030f73b3c401a230c14d89fdd6ac45713f8a5b