Remove "dir=auto" attribute from TypeaheadSearch and Input components
Although the html spec states that the directionality of a
directionality-capable attribute should be determined by the first
character in the attribute's value [1], at least Chrome, Safari, and
Firefox (probably others) don't seem to follow this [2] which can result
in placeholder text appearing on the wrong side of the input (see
T279674).

Because of this, setting the `dir` attribute on input elements
should be up to the client to decide (this is possible via
`v-bind="$attrs"`). Notably, Vector legacy's search input does not set a
dir attribute which is how it bypasses this issue.

[1] https://html.spec.whatwg.org/multipage/dom.html#directionality-of-the-attribute
[2] https://github.com/whatwg/html/issues/4903

Bug: T279674
Change-Id: Ib98f2e7cdc9c78731cdcc7928a4c418693bc252d