TypeaheadSearch: Avoid use of .native event binding
The @mousedown.native binding on the TypeaheadSuggestion component
doesn't play well with the Vue 3 migration build. Instead, have the
TypeaheadSuggestion component emit a mousedown event, and listen to
that.

This was causing a bug when running under the Vue 3 migration build
where clicking a TypeaheadSuggestion didn't do anything, because the
mousedown event was no longer being prevented, so it took focus away
from the input, which closed the menu. By the time the user released the
mouse, the menu was already closed, so no click event was fired.

Bug: T297421
Change-Id: I24a88a862ff30ccdeafb0ae22cea9ee37a6f6ffe